+++ src/applications/medical-copays/sass/medical-copays.scss
@@ -1,2 +1,3 @@
 @import "~@department-of-veterans-affairs/formation/sass/shared-variables";
 @import "./partials/expandable-alert";
+@import "./partials/balance-card";
