+++ src/applications/medical-copays/sass/partials/_balance-card.scss
@@ -1,0 +1,24 @@
+@import "~@department-of-veterans-affairs/formation/sass/shared-variables";
+
+.card-balance {
+  font-family: "Bitter";
+  font-size: 24px;
+}
+
+.card-heading {
+  font-family: "Bitter";
+  font-size: 17px;
+  color: $color-gray-dark;
+}
+
+.card-content {
+  display: flex;
+  align-items: baseline;
+  margin-bottom: 16px;
+  i {
+    padding-right: 10px;
+  }
+  p {
+    margin: 0;
+  }
+}
